cache: lock the repo with a pid file; automatic cleaning
